refactor(#330): Replace TermShared module with methods in TermBase class
